+++ previous-versions/common/src/codatcommon/models/operations/delete_api_key.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
+from ...models.errors import errormessage as errors_errormessage
 from typing import Optional
 
 
@@ -19,11 +19,11 @@ class DeleteAPIKeyRequest:
 class DeleteAPIKeyResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
+    error_message: Optional[errors_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Too many requests were made in a given amount of time. Wait a short period and then try again."""