+++ previous-versions/common/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.34.0] previous-versions/common
 ### Releases
 - [PyPI v0.34.0] https://pypi.org/project/codat-common/0.34.0 - previous-versions/common
+
+## 2023-12-11 01:27:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.35.0] previous-versions/common
+### Releases
+- [PyPI v0.35.0] https://pypi.org/project/codat-common/0.35.0 - previous-versions/common